+++ test/asm-tests.ts
@@ -25,6 +25,7 @@
 import {AsmParser} from '../lib/parsers/asm-parser';
 import {VcAsmParser} from '../lib/parsers/asm-parser-vc';
 import {AsmRegex} from '../lib/parsers/asmregex';
+import { makeFakeParseFiltersAndOutputOptions } from './utils';
 
 describe('ASM CL parser', () => {
     it('should work for error documents', () => {
@@ -45,21 +46,21 @@ describe('ASM CL parser', () => {
 describe('ASM regex base class', () => {
     it('should leave unfiltered lines alone', () => {
         const line = '     this    is    a line';
-        AsmRegex.filterAsmLine(line, {}).should.equal(line);
+        AsmRegex.filterAsmLine(line, makeFakeParseFiltersAndOutputOptions({})).should.equal(line);
     });
     it('should use up internal whitespace when asked', () => {
-        AsmRegex.filterAsmLine('     this    is    a line', {trim: true}).should.equal('  this is a line');
-        AsmRegex.filterAsmLine('this    is    a line', {trim: true}).should.equal('this is a line');
+        AsmRegex.filterAsmLine('     this    is    a line', makeFakeParseFiltersAndOutputOptions({trim: true})).should.equal('  this is a line');
+        AsmRegex.filterAsmLine('this    is    a line', makeFakeParseFiltersAndOutputOptions({trim: true})).should.equal('this is a line');
     });
     it('should keep whitespace in strings', () => {
-        AsmRegex.filterAsmLine('equs     "this    string"', {trim: true}).should.equal('equs "this    string"');
-        AsmRegex.filterAsmLine('     equs     "this    string"', {trim: true}).should.equal('  equs "this    string"');
-        AsmRegex.filterAsmLine('equs     "this    \\"  string  \\""', {trim: true}).should.equal(
+        AsmRegex.filterAsmLine('equs     "this    string"', makeFakeParseFiltersAndOutputOptions({trim: true})).should.equal('equs "this    string"');
+        AsmRegex.filterAsmLine('     equs     "this    string"', makeFakeParseFiltersAndOutputOptions({trim: true})).should.equal('  equs "this    string"');
+        AsmRegex.filterAsmLine('equs     "this    \\"  string  \\""', makeFakeParseFiltersAndOutputOptions({trim: true})).should.equal(
             'equs "this    \\"  string  \\""',
         );
     });
     it('should not get upset by mismatched strings', () => {
-        AsmRegex.filterAsmLine('a   "string    \'yeah', {trim: true}).should.equal('a "string \'yeah');
+        AsmRegex.filterAsmLine('a   "string    \'yeah', makeFakeParseFiltersAndOutputOptions({trim: true})).should.equal('a "string \'yeah');
     });
 });
