+++ test/nim-tests.ts
@@ -26,7 +26,8 @@ import path from 'path';
 
 import {NimCompiler} from '../lib/compilers/nim';
 
-import {makeCompilationEnvironment, should} from './utils';
+import {makeCompilationEnvironment, makeFakeCompilerInfo, should} from './utils';
+import { unwrap } from '../lib/assert';
 
 const languages = {
     nim: {id: 'nim'},
@@ -35,7 +36,7 @@ const languages = {
 describe('Nim', () => {
     let ce;
     const info = {
-        exe: null,
+        exe: "/dev/null",
         remote: true,
         lang: languages.nim.id,
     };
@@ -45,20 +46,20 @@ describe('Nim', () => {
     });
 
     it('Nim should not allow --run/-r parameter', () => {
-        const compiler = new NimCompiler(info, ce);
+        const compiler = new NimCompiler(makeFakeCompilerInfo(info), ce);
         compiler.filterUserOptions(['c', '--run', '--something']).should.deep.equal(['c', '--something']);
         compiler.filterUserOptions(['cpp', '-r', '--something']).should.deep.equal(['cpp', '--something']);
     });
 
     it('Nim compile to Cpp if not asked otherwise', () => {
-        const compiler = new NimCompiler(info, ce);
+        const compiler = new NimCompiler(makeFakeCompilerInfo(info), ce);
         compiler.filterUserOptions([]).should.deep.equal(['compile']);
         compiler.filterUserOptions(['badoption']).should.deep.equal(['compile', 'badoption']);
         compiler.filterUserOptions(['js']).should.deep.equal(['js']);
     });
 
     it('test getCacheFile from possible user-options', () => {
-        const compiler = new NimCompiler(info, ce),
+        const compiler = new NimCompiler(makeFakeCompilerInfo(info), ce),
             input = 'test.min',
             folder = path.join('/', 'tmp/'),
             expected = {
@@ -68,7 +69,7 @@ describe('Nim', () => {
             };
 
         for (const lang of ['cpp', 'c', 'objc']) {
-            compiler.getCacheFile([lang], input, folder).should.equal(expected[lang]);
+            unwrap(compiler.getCacheFile([lang], input, folder)).should.equal(expected[lang]);
         }
 
         should.equal(compiler.getCacheFile([], input, folder), null);
