+++ test/options-handler.ts
